Parts of script were commented out; some features require specific actions to be done
The DisableBoolAnimator and NavigateTo methods have been commented out completely, as syntax errors will be produced.  The same applies with the ExitGame method's structure, and not what's in that method.  When features get added, like the necessary animations and scene management, then these parts of the code can be uncommented.